+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ### v0.0.4
 
+Added support for custom props.
+
 Added new props:
 
 - cursor