+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this project will be documented in this file.
 
+### v0.0.5
+
+Added explicit CustomProp type and make CustomProps a slice to preserve ordering.
+
 ### v0.0.4
 
 Added support for custom props.